debug

+++ part1/setInCache.c
@@ -45,7 +45,7 @@ void setBit(cache_t* cache, uint64_t location, uint8_t value) {
     if(shiftAmount == 0)
         cache->contents[byteLoc] = value<<8;
     else{
-        uint32_t temp = cache->contents[byteLoc];
+        uint8_t temp = cache->contents[byteLoc];
         temp = (((temp>>(8-shiftAmount+1))<<1)|value)<<(8-shiftAmount);
         cache->contents[byteLoc] = (uint8_t) temp;
     }
